mysite/requirements.txt
anyio==3.7.0
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.2
django-crispy-forms==2.0
djangorestframework==3.14.0
easy-install==66.0.2
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
pytz==2023.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
tailwind==3.1.5b0
traitlets==5.9.0
tzdata==2023.3
watchfiles==0.19.0
wcwidth==0.2.6
